#1: added initial start page of aurelia frontend
Added the initial starting page of the aurelia based frontend. This includes the top and side navigation, a main content area, and a footer. The content area has a basic news service implemented which will show the last news. Loading bar is based on nprogress and will indicate requesting/navigating