#1: rework scss files and navigation design
Some scss files are split up in multiple to reflect their use. The
layout scss is now only handling positioning on the page and has no more
styling effects. Other files are there now for each part of the layout.
The navigation was completely reworked in design and in scss usability.
The hover effect menu still needs some work.
Introduced a new card-default class for showing cards with a default
background-color and border. This is added to all currently existing
cards that do not have a specific coloring to have a contrast to the
content background.