[#5] user: add REST resource for user registration
Added the REST resource for the user registration on the front end.
Validation is the same as for the front end. If an error occurs,
validation errors are returned as json. Currently, the notification only
shows a message indicating that there was an error. The errors are
normally listed within the input field, but should be added to the
notification box in the future.
In addition, the CSRF middleware was removed as ajax request currently
do not include the toke in requests. This has to be adjusted in the
future.

---

 app/Http/Controllers/UserController.php        | 20 +++++++++++++++++++
 app/Http/Kernel.php                            |  2 +-
 app/Http/routes.php                            |  3 +++
 .../assets/app/services/users/user-service.js  |  6 +++++-
 .../assets/app/view-models/users/register.js   | 20 +++++++++----------
 5 files changed, 39 insertions(+), 12 deletions(-)