[#7] database: cleanup database migrations files
The inital migration files are shipped with laravel and are not needed
for the database structure. Future database changes will result in new
migration files. The first two tables are added as a migration file and
need to be used to set up the database.
In addition, the database name was updated for a clean setup.

---

 config/database.php                            |  2 +-
 .../2014_10_12_000000_create_users_table.php   | 34 -------------------
 ..._12_100000_create_password_resets_table.php | 31 -----------------
 .../2015_09_20_111406_create_news_table.php    | 33 ++++++++++++++++++
 .../2015_09_20_112452_create_users_table.php   | 33 ++++++++++++++++++
 5 files changed, 67 insertions(+), 66 deletions(-)