[#5] user: add debounce to username and mail check
Username and email validation is triggered each type the user types in a
new character to the input fields. This causes high load on the server.

A debounce is added to the validation to reduce the amount of request
sent to the server. This is currently set to 150ms for both, email and
username.

---

 resources/assets/app/view-models/users/register.js | 5 ++---
 1 file changed, 2 insertions(+), 3 deletions(-)