[#10] pages: hypenate views and show title correct
The pages view-module is loading views dynamically based on their
hyphenated name. The hyphenate function of aurelia-templating was
renamed with the latest release. In addition, the title of each page
view is showing the same.

The hyphenate function was renamed to fit the aurelia-templating
package export. In addition, each view now gets its name in addition to
the application title. This is done by using the url param and titlecase
it. A corresponding funtion is now available via util.js called
_titlecase.

---

 resources/assets/app/utilities/util.js          | 9 +++++++++
 resources/assets/app/view-models/pages/index.js | 8 +++++---
 2 files changed, 14 insertions(+), 3 deletions(-)