[#1] style: change bg color and add bg content var
The background of the page is colored with one variable. A distinction
between page background and main content background is not possible.

A new variable is introduced which allows to specify both background
colors independent from the other one. In addition, the page background
is colored white now to fit to the footer color. The main content is
colored as before.

---

 resources/assets/css/include/_layout.scss    |  2 +-
 resources/assets/css/include/_page.scss      |  1 +
 resources/assets/css/include/_variables.scss | 25 ++++++++++++---------
 3 files changed, 16 insertions(+), 12 deletions(-)