[#6] auth: implement sign in route in auth controller
Added the auth controller to application for authentication based on JWT
using the laravel JWT plugin. The corresponding routes are added to the
applications route configuration.

The JWT secret is added to the env example file.

The user model was listing the wrong field names as fillable.
The username field is named correct now.

---

 .env.example                            |  2 ++
 app/Http/Controllers/AuthController.php | 27 ++++++++++++++++++++++++++
 app/Http/routes.php                     |  2 ++
 app/Models/User.php                     |  2 +-
 4 files changed, 32 insertions(+), 1 deletion(-)