[#6] auth: refactor and update to latest JWT version
Updated aurelia-auth to latest release and fix some issues regarding the
usage of the http client (injected the fetch client config instead of
the http client config as before).

Moved the view6view-model for signing in to an own package auth.
Related files are updated according to new package name aurelia-auth and
routes are adjusted to the new package.

The sign up functionality itself is implemented as well.

---

 package.json                                   |   2 +-
 public/config.js                               | 191 ++++++------------
 .../assets/app/configuration/auth-config.js    |   8 +-
 .../assets/app/configuration/legendsrising.js  |   2 +-
 .../assets/app/configuration/router-config.js  |   4 +-
 .../app/templates/elements/navigation-side.js  |   2 +-
 .../app/templates/elements/navigation-top.html |  16 +-
 .../app/templates/elements/navigation-top.js   |  27 +++
 resources/assets/app/view-models/app.js        |  10 +-
 .../assets/app/view-models/auth/signin.js      |  24 +++
 .../assets/app/view-models/users/signin.js     |   3 -
 resources/assets/app/views/auth/signin.html    |  34 ++++
 resources/assets/app/views/users/signin.html   |  43 ----
 13 files changed, 171 insertions(+), 195 deletions(-)