fix(auth): fix sign in and out redirect
The sign in and sign out process is reloading the complete application
instead of only redirecting to the specified route.

The aurelia-auth plugin has a bug when using aurelias push state API. A
work around which specifies no routes for redirection in the auth-config
avoids redirection. This redirection happens directly in the VMs with
aurelias router navigation.

---

 resources/assets/app/configuration/auth-config.js |  4 ++--
 resources/assets/app/view-models/auth/signin.js   | 12 ++++++++----
 resources/assets/app/view-models/auth/signout.js  | 12 ++++++++----
 resources/assets/app/views/auth/signout.html      |  3 +++
 4 files changed, 21 insertions(+), 10 deletions(-)